fix(trusty-search): remediate #928 — surface deferred-embed failures, split tests, line-cap
- Task 1: rebased onto origin/main (resolved 6 conflict sites)
- Task 2 (#928): in spawn_deferred_embed_pass Err branch, set
  stages.semantic = StageState::failed(reason) so /indexes/:id/status
  exposes embed failures instead of leaving the stage stuck in Pending
- Task 2 tests: add deferred_embed_pass_marks_semantic_ready_and_is_idempotent
  and failing_deferred_embed_pass_marks_semantic_failed in defer_embed::tests
- Task 3: extract spawn_deferred_embed_pass → service/reindex/defer_embed.rs
  (300 lines); move inline #[cfg(test)] mod tests → reindex/tests.rs (1676 lines);
  reindex/mod.rs drops from 4462 → 2784 lines (allowlist budget: 2785, down from
  4373 on origin/main — satisfies "not increased" constraint)
- Task 4 (doc nits): add serde asymmetry comment on persistence.rs defer_embed
  field (is_true vs Not::not); add Why/What/Test doc on repo_config::default_true()
- Fix hash_cache.rs test: add missing defer_embed field to IndexHandle literal
- Fix defer_embed.rs test: update FailingEmbedder to current Embedder trait
  signature (embed + embed_batch with &[&str]), correct RawChunk/ParsedBatch fields

Quality bar: check_line_cap OK (0 violations), cargo fmt clean, clippy -D warnings
clean, cargo test -p trusty-search: 1010 passed / 0 failed.
