+++ .line-cap-allowlist.tsv
@@ -139,22 +139,22 @@ crates/trusty-review/src/pipeline/runner_tests.rs	730
 crates/trusty-review/src/pipeline/verify_tests.rs	533
 crates/trusty-search/src/commands/doctor_checks.rs	612
 crates/trusty-search/src/commands/integrate.rs	688
-crates/trusty-search/src/commands/reindex_engine.rs	1791
+crates/trusty-search/src/commands/reindex_engine.rs	1817
 crates/trusty-search/src/commands/reindex_ui.rs	1429
-crates/trusty-search/src/commands/start.rs	2194
+crates/trusty-search/src/commands/start.rs	2197
 crates/trusty-search/src/core/chunker/mod.rs	1883
 crates/trusty-search/src/core/classifier.rs	1032
 crates/trusty-search/src/core/corpus.rs	1685
 crates/trusty-search/src/core/indexer/docs_penalty.rs	543
-crates/trusty-search/src/core/indexer/ingest.rs	1186
+crates/trusty-search/src/core/indexer/ingest.rs	1214
 crates/trusty-search/src/core/indexer/mod.rs	1307
 crates/trusty-search/src/core/indexer/persist.rs	740
 crates/trusty-search/src/core/indexer/search.rs	1109
 crates/trusty-search/src/core/indexer/tests.rs	3426
 crates/trusty-search/src/core/memory_policy.rs	1264
 crates/trusty-search/src/core/migration/mod.rs	649
-crates/trusty-search/src/core/registry.rs	734
-crates/trusty-search/src/core/repo_config.rs	586
+crates/trusty-search/src/core/registry.rs	732
+crates/trusty-search/src/core/repo_config.rs	595
 crates/trusty-search/src/core/search/hierarchy.rs	658
 crates/trusty-search/src/core/store.rs	1275
 crates/trusty-search/src/core/symbol_graph.rs	1667
@@ -164,8 +164,9 @@ crates/trusty-search/src/service/call_chain.rs	954
 crates/trusty-search/src/service/daemon.rs	774
 crates/trusty-search/src/service/embedder_supervisor.rs	1020
 crates/trusty-search/src/service/grep.rs	827
-crates/trusty-search/src/service/persistence.rs	1014
-crates/trusty-search/src/service/reindex/mod.rs	3820
+crates/trusty-search/src/service/persistence.rs	1017
+crates/trusty-search/src/service/reindex/mod.rs	2784
+crates/trusty-search/src/service/reindex/tests.rs	1674
 crates/trusty-search/src/service/walker.rs	1344
 crates/trusty-search/src/service/warm_boot/probe.rs	518
 crates/trusty-search/tests/baseline_trusty_tools.rs	721
